Move utilities that are only used to resolve an `AnnotationSpec` for PSI elements into a separate file.
