+++ compiler-utils/src/main/java/com/squareup/anvil/compiler/internal/KotlinPoetUtils.kt
@@ -5,13 +5,10 @@ package com.squareup.anvil.compiler.internal
 import com.squareup.anvil.annotations.ExperimentalAnvilApi
 import com.squareup.anvil.compiler.api.AnvilCompilationException
 import com.squareup.anvil.compiler.internal.reference.canResolveFqName
-import com.squareup.anvil.compiler.internal.reference.toClassReferencePsiOrNull
 import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.ClassName
-import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.LambdaTypeName
-import com.squareup.kotlinpoet.MemberName
 import com.squareup.kotlinpoet.ParameterSpec
 import com.squareup.kotlinpoet.ParameterizedTypeName
 import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
@@ -27,32 +24,17 @@ import org.jetbrains.kotlin.descriptors.ClassDescriptor
 import org.jetbrains.kotlin.descriptors.ModuleDescriptor
 import org.jetbrains.kotlin.descriptors.PackageFragmentDescriptor
 import org.jetbrains.kotlin.descriptors.PropertyDescriptor
-import org.jetbrains.kotlin.fir.lightTree.converter.nameAsSafeName
 import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.name.FqName
-import org.jetbrains.kotlin.psi.KtAnnotationEntry
 import org.jetbrains.kotlin.psi.KtCallableDeclaration
-import org.jetbrains.kotlin.psi.KtClass
-import org.jetbrains.kotlin.psi.KtClassLiteralExpression
 import org.jetbrains.kotlin.psi.KtClassOrObject
-import org.jetbrains.kotlin.psi.KtCollectionLiteralExpression
-import org.jetbrains.kotlin.psi.KtElement
-import org.jetbrains.kotlin.psi.KtExpression
-import org.jetbrains.kotlin.psi.KtFile
 import org.jetbrains.kotlin.psi.KtFunctionType
-import org.jetbrains.kotlin.psi.KtNameReferenceExpression
-import org.jetbrains.kotlin.psi.KtNamedFunction
 import org.jetbrains.kotlin.psi.KtNullableType
-import org.jetbrains.kotlin.psi.KtObjectDeclaration
 import org.jetbrains.kotlin.psi.KtProjectionKind
-import org.jetbrains.kotlin.psi.KtProperty
-import org.jetbrains.kotlin.psi.KtQualifiedExpression
 import org.jetbrains.kotlin.psi.KtTypeElement
 import org.jetbrains.kotlin.psi.KtTypeReference
 import org.jetbrains.kotlin.psi.KtUserType
-import org.jetbrains.kotlin.psi.KtValueArgument
-import org.jetbrains.kotlin.psi.psiUtil.containingClass
 import org.jetbrains.kotlin.psi.psiUtil.parentsWithSelf
 import org.jetbrains.kotlin.resolve.descriptorUtil.parents
 import org.jetbrains.kotlin.resolve.descriptorUtil.parentsWithSelf
@@ -92,7 +74,7 @@ public fun FqName.asClassName(module: ModuleDescriptor): ClassName {
     ?: throw AnvilCompilationException("Couldn't parse ClassName for $this.")
 }
 
-private fun FqName.asClassNameOrNull(module: ModuleDescriptor): ClassName? {
+internal fun FqName.asClassNameOrNull(module: ModuleDescriptor): ClassName? {
   val segments = pathSegments().map { it.asString() }
 
   // If the first sentence case is not the last segment of the path it becomes ambiguous,
@@ -119,26 +101,6 @@ private fun FqName.asClassNameOrNull(module: ModuleDescriptor): ClassName? {
 
   // No matching class found. It's either a package name only or doesn't exist.
   return null
-}
-
-@ExperimentalAnvilApi
-public fun FqName.asMemberName(module: ModuleDescriptor): MemberName {
-  val segments = pathSegments().map { it.asString() }
-  val simpleName = segments.last()
-  val prefixFqName = FqName.fromSegments(segments.dropLast(1))
-  return prefixFqName.asClassNameOrNull(module)?.let {
-    val classOrObj = prefixFqName.toClassReferencePsiOrNull(module)?.clazz
-    val imported = if (classOrObj is KtClass) {
-      // Must be in a companion, check its name
-      // We do this because accessors could be just Foo.CONSTANT but to member import it we need
-      // to include the companion path
-      val companionName = classOrObj.companionObjects.first().name ?: "Companion"
-      it.nestedClass(companionName)
-    } else {
-      it
-    }
-    MemberName(imported, simpleName)
-  } ?: MemberName(prefixFqName.pathSegments().joinToString("."), simpleName)
 }
 
 @ExperimentalAnvilApi
@@ -373,113 +335,3 @@ public fun FileSpec.Companion.buildFile(
     .addComment(generatorComment)
     .build()
     .writeToString()
-
-/** Returns an [AnnotationSpec] representation of this [KtAnnotationEntry]. */
-@ExperimentalAnvilApi
-@Suppress("DeprecatedCallableAddReplaceWith")
-@Deprecated(
-  "Don't rely on PSI and make the code agnostic to the underlying implementation."
-)
-public fun KtAnnotationEntry.toAnnotationSpec(
-  module: ModuleDescriptor
-): AnnotationSpec {
-  return AnnotationSpec.builder(requireFqName(module).asClassName(module))
-    .apply {
-      valueArguments
-        .filterIsInstance<KtValueArgument>()
-        .mapNotNull { valueArgument ->
-          valueArgument.getArgumentExpression()?.codeBlock(module)
-        }.forEach {
-          addMember(it)
-        }
-    }
-    .build()
-}
-
-internal fun KtExpression.codeBlock(module: ModuleDescriptor): CodeBlock {
-  return when (this) {
-    // MyClass::class
-    is KtClassLiteralExpression -> {
-      val className = requireFqName(module).asClassName(module)
-      CodeBlock.of("%T::class", className)
-    }
-    // enums or qualified references
-    is KtNameReferenceExpression, is KtQualifiedExpression -> {
-      val fqName = try {
-        requireFqName(module)
-      } catch (e: AnvilCompilationException) {
-        if (this is KtNameReferenceExpression) {
-          null
-        } else {
-          throw e
-        }
-      }
-      if (fqName != null) {
-        if (fqName.canResolveFqName(module)) {
-          CodeBlock.of("%T", fqName.asClassName(module))
-        } else {
-          CodeBlock.of("%M", fqName.asMemberName(module))
-        }
-      } else {
-        // It's (hopefully) an in-scope constant, look up the hierarchy.
-        val ref = (this as KtNameReferenceExpression).getReferencedName()
-        findConstantDefinitionInScope(ref)
-          ?.asMemberName(module)
-          ?.let { memberRef ->
-            CodeBlock.of("%M", memberRef)
-          }
-          // If we reach here, it's a top-level constant defined in the same package but a different
-          // file. In this case, we can just do the same in our generated code because we're in the
-          // same package and can play by the same rules.
-          ?: CodeBlock.of("%L", ref)
-      }
-    }
-    is KtCollectionLiteralExpression -> CodeBlock.of(
-      getInnerExpressions()
-        .map { it.codeBlock(module) }
-        .joinToString(prefix = "[", postfix = "]")
-    )
-    // literals
-    else -> CodeBlock.of("%L", text)
-  }
-}
-
-private fun List<KtProperty>.containsConstPropertyWithName(name: String): Boolean {
-  return any { property ->
-    property.hasModifier(KtTokens.CONST_KEYWORD) && property.name == name
-  }
-}
-
-private fun PsiElement.findConstantDefinitionInScope(name: String): FqName? {
-  when (this) {
-    is KtProperty, is KtNamedFunction -> {
-      // Function or Property, traverse up
-      return (this as KtElement).containingClass()?.findConstantDefinitionInScope(name)
-        ?: containingFile.findConstantDefinitionInScope(name)
-    }
-    is KtObjectDeclaration -> {
-      if (properties(includeCompanionObjects = false).containsConstPropertyWithName(name)) {
-        return requireFqName().child(name.nameAsSafeName())
-      } else if (isCompanion()) {
-        // Nowhere else to look and don't try to traverse up because this is only looked at from a
-        // class already.
-        return null
-      }
-    }
-    is KtFile -> {
-      if (children.filterIsInstance<KtProperty>().containsConstPropertyWithName(name)) {
-        return packageFqName.child(name.nameAsSafeName())
-      }
-    }
-    is KtClass -> {
-      // Look in companion object or traverse up
-      return companionObjects.asSequence()
-        .mapNotNull { it.findConstantDefinitionInScope(name) }
-        .firstOrNull()
-        ?: containingClass()?.findConstantDefinitionInScope(name)
-        ?: containingKtFile.findConstantDefinitionInScope(name)
-    }
-  }
-
-  return parent?.findConstantDefinitionInScope(name)
-}
